Unified theme for all Sign in buttons
Um, so I made all the Sign in buttons look the same now - same like styling and functiounality across all of them.
Also, the button that used to send people to /login (which doesn’t even exist .-. ) now opens the sign-in popup instead.
Did the same for the one on the coinSymbol page so no more redirect to "/", it just opens the popup now.
Oh, and the comments section button also opens the Popup. Just annoyed me that they looked differently and redirected to 404 Pages.